Codigo/05-DeployModelos/requirements.txt
absl-py==1.2.0
aiobotocore==2.3.4
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
botocore==1.24.21
cachetools==5.2.0
certifi==2022.6.15
cftime==1.6.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
configobj==5.0.6
Flask==2.2.2
flatbuffers==1.12
frozenlist==1.3.1
fsspec==2022.7.1
gast==0.4.0
GOES==3.3
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
imageio==2.28.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
lazy_loader==0.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
netCDF4==1.6.0
networkx==3.1
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
Pillow==9.5.0
plotly==5.14.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pykdtree==1.3.5
pyparsing==3.0.9
pyproj==3.3.1
pyresample==1.25.1
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3fs==2022.7.1
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
tenacity==8.2.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2023.4.12
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
